requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
et-xmlfile==2.0.0
idna==3.10
numpy==2.2.2
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
tweepy==4.15.0
tzdata==2024.2
urllib3==2.3.0
